added modules for codeception.yml (8)
