--- tests/SprykerEcoTest/Zed/Easycredit/_support/ModuleNameTester.php
@@ -1,26 +1,0 @@
-<?php
-namespace SprykerEcoTest\Zed\ModuleName;
-
-/**
- * Inherited Methods
- * @method void wantToTest($text)
- * @method void wantTo($text)
- * @method void execute($callable)
- * @method void expectTo($prediction)
- * @method void expect($prediction)
- * @method void amGoingTo($argumentation)
- * @method void am($role)
- * @method void lookForwardTo($achieveValue)
- * @method void comment($description)
- * @method void pause()
- *
- * @SuppressWarnings(PHPMD)
-*/
-class ModuleNameTester extends \Codeception\Actor
-{
-    use _generated\ModuleNameTesterActions;
-
-    /**
-     * Define custom actions here
-     */
-}
