+++ tests/SprykerEcoTest/Zed/Easycredit/_support/_generated/EasycreditTesterActions.php
@@ -1,4 +1,4 @@
-<?php  //[STAMP] 5149d604bd29fbf4f8ca27c7a4f09117
+<?php  //[STAMP] c5b0a019996ce1a8e630957e24485789
 namespace SprykerEcoTest\Zed\Easycredit\_generated;
 
 // This class was automatically generated by build task
@@ -2052,11 +2052,251 @@ trait EasycreditTesterActions
     /**
      * [!] Method is generated. Documentation taken from corresponding module.
      *
-     * @return int
-     * @see \SprykerTest\Shared\Sales\Helper\SalesHelper::createOrder()
+     * @param string $key
+     * @param mixed $value
+     * @param string|null $onlyFor
+     *
+     * @return void
+     * @see \SprykerTest\Shared\Testify\Helper\DependencyHelper::setDependency()
      */
-    public function createOrder(): int {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('createOrder', func_get_args()));
+    public function setDependency(string $key, $value, ?string $onlyFor = NULL): void {
+        $this->getScenario()->runStep(new \Codeception\Step\Action('setDependency', func_get_args()));
+    }
+
+ 
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * @param string|null $moduleName
+     *
+     * @throws \Exception
+     *
+     * @return \Spryker\Zed\Kernel\Container
+     * @see \SprykerTest\Zed\Testify\Helper\AbstractDependencyProviderHelper::getModuleContainer()
+     */
+    public function getModuleContainer(?string $moduleName = NULL): \Spryker\Zed\Kernel\Container {
+        return $this->getScenario()->runStep(new \Codeception\Step\Action('getModuleContainer', func_get_args()));
+    }
+
+ 
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * @param string $methodName
+     * @param mixed $return
+     * @param string|null $moduleName
+     *
+     * @throws \Exception
+     *
+     * @return \Spryker\Zed\Kernel\AbstractBundleDependencyProvider
+     * @see \SprykerTest\Zed\Testify\Helper\AbstractDependencyProviderHelper::mockDependencyProviderMethod()
+     */
+    public function mockDependencyProviderMethod(string $methodName, $return, ?string $moduleName = NULL): \Spryker\Zed\Kernel\AbstractBundleDependencyProvider {
+        return $this->getScenario()->runStep(new \Codeception\Step\Action('mockDependencyProviderMethod', func_get_args()));
+    }
+
+ 
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * @param string $key
+     * @param array|string|float|int|bool $value
+     *
+     * @return void
+     * @see \SprykerTest\Shared\Testify\Helper\LocatorHelper::setConfig()
+     */
+    public function setConfig(string $key, $value): void {
+        $this->getScenario()->runStep(new \Codeception\Step\Action('setConfig', func_get_args()));
+    }
+
+ 
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * @param string $key
+     * @param array|string|float|int|bool $value
+     *
+     * @return void
+     * @see \SprykerTest\Shared\Testify\Helper\ConfigHelper::mockEnvironmentConfig()
+     */
+    public function mockEnvironmentConfig(string $key, $value): void {
+        $this->getScenario()->runStep(new \Codeception\Step\Action('mockEnvironmentConfig', func_get_args()));
+    }
+
+ 
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * @param string $methodName
+     * @param mixed $return
+     * @param string|null $moduleName
+     *
+     * @throws \Exception
+     *
+     * @return \Spryker\Shared\Kernel\AbstractBundleConfig|null
+     * @see \SprykerTest\Shared\Testify\Helper\ConfigHelper::mockConfigMethod()
+     */
+    public function mockConfigMethod(string $methodName, $return, ?string $moduleName = NULL): ?\Spryker\Shared\Kernel\AbstractBundleConfig {
+        return $this->getScenario()->runStep(new \Codeception\Step\Action('mockConfigMethod', func_get_args()));
+    }
+
+ 
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * @param string $methodName
+     * @param mixed $return
+     * @param string|null $moduleName
+     *
+     * @throws \Exception
+     *
+     * @return \Spryker\Shared\Kernel\AbstractSharedConfig|null
+     * @see \SprykerTest\Shared\Testify\Helper\ConfigHelper::mockSharedConfigMethod()
+     */
+    public function mockSharedConfigMethod(string $methodName, $return, ?string $moduleName = NULL): ?\Spryker\Shared\Kernel\AbstractSharedConfig {
+        return $this->getScenario()->runStep(new \Codeception\Step\Action('mockSharedConfigMethod', func_get_args()));
+    }
+
+ 
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * @param string|null $moduleName
+     *
+     * @return \Spryker\Shared\Kernel\AbstractBundleConfig
+     * @see \SprykerTest\Shared\Testify\Helper\ConfigHelper::getModuleConfig()
+     */
+    public function getModuleConfig(?string $moduleName = NULL): \Spryker\Shared\Kernel\AbstractBundleConfig {
+        return $this->getScenario()->runStep(new \Codeception\Step\Action('getModuleConfig', func_get_args()));
+    }
+
+ 
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * @param string|null $moduleName
+     *
+     * @return \Spryker\Shared\Kernel\AbstractSharedConfig|null
+     * @see \SprykerTest\Shared\Testify\Helper\ConfigHelper::getSharedModuleConfig()
+     */
+    public function getSharedModuleConfig(?string $moduleName = NULL): ?\Spryker\Shared\Kernel\AbstractSharedConfig {
+        return $this->getScenario()->runStep(new \Codeception\Step\Action('getSharedModuleConfig', func_get_args()));
+    }
+
+ 
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * @param string $moduleName
+     *
+     * @return bool
+     * @see \SprykerTest\Shared\Testify\Helper\ConfigHelper::configExists()
+     */
+    public function configExists(string $moduleName): bool {
+        return $this->getScenario()->runStep(new \Codeception\Step\Action('configExists', func_get_args()));
+    }
+
+ 
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * @param string $key
+     *
+     * @return void
+     * @see \SprykerTest\Shared\Testify\Helper\ConfigHelper::removeConfig()
+     */
+    public function removeConfig(string $key): void {
+        $this->getScenario()->runStep(new \Codeception\Step\Action('removeConfig', func_get_args()));
+    }
+
+ 
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * @param string $methodName
+     * @param mixed $return
+     * @param string|null $moduleName
+     *
+     * @throws \Exception
+     *
+     * @return \Spryker\Zed\Kernel\Business\AbstractFacade
+     * @see \SprykerTest\Zed\Testify\Helper\Business\BusinessHelper::mockFacadeMethod()
+     */
+    public function mockFacadeMethod(string $methodName, $return, ?string $moduleName = NULL) {
+        return $this->getScenario()->runStep(new \Codeception\Step\Action('mockFacadeMethod', func_get_args()));
+    }
+
+ 
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * @deprecated Use {@link \SprykerTest\Zed\Testify\Helper\Business\BusinessHelper::getFacade()} instead.
+     *
+     * @return \Spryker\Zed\Kernel\Business\AbstractFacade
+     * @see \SprykerTest\Shared\Testify\Helper\LocatorHelper::getFacade()
+     */
+    public function getFacade(): \Spryker\Zed\Kernel\Business\AbstractFacade {
+        return $this->getScenario()->runStep(new \Codeception\Step\Action('getFacade', func_get_args()));
+    }
+
+ 
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * @param string $methodName
+     * @param mixed $return
+     * @param string|null $moduleName
+     *
+     * @throws \Exception
+     *
+     * @return \Spryker\Zed\Kernel\Business\AbstractBusinessFactory|object
+     * @see \SprykerTest\Zed\Testify\Helper\Business\BusinessHelper::mockFactoryMethod()
+     */
+    public function mockFactoryMethod(string $methodName, $return, ?string $moduleName = NULL) {
+        return $this->getScenario()->runStep(new \Codeception\Step\Action('mockFactoryMethod', func_get_args()));
+    }
+
+ 
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * @param string $methodName
+     * @param mixed $return
+     * @param string|null $moduleName
+     *
+     * @throws \Exception
+     *
+     * @return \Spryker\Zed\Kernel\Business\AbstractBusinessFactory|object
+     * @see \SprykerTest\Zed\Testify\Helper\Business\BusinessHelper::mockSharedFactoryMethod()
+     */
+    public function mockSharedFactoryMethod(string $methodName, $return, ?string $moduleName = NULL) {
+        return $this->getScenario()->runStep(new \Codeception\Step\Action('mockSharedFactoryMethod', func_get_args()));
+    }
+
+ 
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * @param string|null $moduleName
+     *
+     * @return \Spryker\Zed\Kernel\Business\AbstractBusinessFactory
+     * @see \SprykerTest\Zed\Testify\Helper\Business\BusinessHelper::getFactory()
+     */
+    public function getFactory(?string $moduleName = NULL): \Spryker\Zed\Kernel\Business\AbstractBusinessFactory {
+        return $this->getScenario()->runStep(new \Codeception\Step\Action('getFactory', func_get_args()));
+    }
+
+ 
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * @param \Closure $closure
+     *
+     * @return void
+     * @see \SprykerTest\Shared\Testify\Helper\DataCleanupHelper::addCleanup()
+     */
+    public function addCleanup(\Closure $closure): void {
+        $this->getScenario()->runStep(new \Codeception\Step\Action('addCleanup', func_get_args()));
     }
 
  
@@ -2064,37 +2304,85 @@ trait EasycreditTesterActions
      * [!] Method is generated. Documentation taken from corresponding module.
      *
      * @return void
-     * @see \SprykerTest\Shared\Sales\Helper\SalesHelper::createOrderWithOneItem()
+     * @see \SprykerTest\Shared\Testify\Helper\DependencyHelper::clearFactoryContainerCache()
      */
-    public function createOrderWithOneItem(): void {
-        $this->getScenario()->runStep(new \Codeception\Step\Action('createOrderWithOneItem', func_get_args()));
+    public function clearFactoryContainerCache(): void {
+        $this->getScenario()->runStep(new \Codeception\Step\Action('clearFactoryContainerCache', func_get_args()));
     }
 
  
     /**
      * [!] Method is generated. Documentation taken from corresponding module.
      *
-     * @param int $idSalesOrder
-     * @param array $salesOrderItem
+     * Sets a class instance into the Locator cache to ensure the mocked instance is returned when
+     * `$locator->moduleName()->type()` is used.
      *
-     * @return \Orm\Zed\Sales\Persistence\SpySalesOrderItem
-     * @see \SprykerTest\Shared\Sales\Helper\SalesHelper::createSalesOrderItemForOrder()
-     */
-    public function createSalesOrderItemForOrder(int $idSalesOrder, array $salesOrderItem = []): \Orm\Zed\Sales\Persistence\SpySalesOrderItem {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('createSalesOrderItemForOrder', func_get_args()));
-    }
-
- 
-    /**
-     * [!] Method is generated. Documentation taken from corresponding module.
+     * !!! When this method is used the locator will not re-initialize classes with `new` but will return
+     * always the already resolved instances. This can have but should not have side-effects.
      *
-     * @param int $idSalesOrderItem
-     * @param array $discount
+     * @param string $cacheKey
+     * @param mixed $classInstance
      *
      * @return void
-     * @see \SprykerTest\Shared\Sales\Helper\SalesHelper::createDiscountForSalesOrderItem()
+     * @see \SprykerTest\Shared\Testify\Helper\LocatorHelper::addToLocatorCache()
      */
-    public function createDiscountForSalesOrderItem(int $idSalesOrderItem, array $discount = []): void {
-        $this->getScenario()->runStep(new \Codeception\Step\Action('createDiscountForSalesOrderItem', func_get_args()));
+    public function addToLocatorCache(string $cacheKey, $classInstance): void {
+        $this->getScenario()->runStep(new \Codeception\Step\Action('addToLocatorCache', func_get_args()));
+    }
+
+ 
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * @return \Spryker\Shared\Kernel\LocatorLocatorInterface&\Generated\Zed\Ide\AutoCompletion&\Generated\Service\Ide\AutoCompletion&\Generated\Glue\Ide\AutoCompletion
+     * @see \SprykerTest\Shared\Testify\Helper\LocatorHelper::getLocator()
+     */
+    public function getLocator() {
+        return $this->getScenario()->runStep(new \Codeception\Step\Action('getLocator', func_get_args()));
+    }
+
+ 
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * @param array $override
+     * @param string|null $stateMachineProcessName
+     * @param array<\Spryker\Zed\CheckoutExtension\Dependency\Plugin\CheckoutDoSaveOrderInterface> $saveOrderStack
+     *
+     * @return \Generated\Shared\Transfer\SaveOrderTransfer
+     * @see \SprykerTest\Shared\Sales\Helper\SalesDataHelper::haveOrder()
+     */
+    public function haveOrder(array $override = [], ?string $stateMachineProcessName = NULL, array $saveOrderStack = []): \Generated\Shared\Transfer\SaveOrderTransfer {
+        return $this->getScenario()->runStep(new \Codeception\Step\Action('haveOrder', func_get_args()));
+    }
+
+ 
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * @param \Generated\Shared\Transfer\QuoteTransfer $quoteTransfer
+     * @param string|null $stateMachineProcessName
+     * @param array<\Spryker\Zed\CheckoutExtension\Dependency\Plugin\CheckoutDoSaveOrderInterface> $checkoutDoSaveOrderPlugins
+     *
+     * @return \Generated\Shared\Transfer\SaveOrderTransfer
+     * @see \SprykerTest\Shared\Sales\Helper\SalesDataHelper::haveOrderFromQuote()
+     */
+    public function haveOrderFromQuote(\Generated\Shared\Transfer\QuoteTransfer $quoteTransfer, ?string $stateMachineProcessName = NULL, array $checkoutDoSaveOrderPlugins = []): \Generated\Shared\Transfer\SaveOrderTransfer {
+        return $this->getScenario()->runStep(new \Codeception\Step\Action('haveOrderFromQuote', func_get_args()));
+    }
+
+ 
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * @param \Generated\Shared\Transfer\QuoteTransfer $quoteTransfer
+     * @param string|null $stateMachineProcessName
+     * @param array<\Spryker\Zed\CheckoutExtension\Dependency\Plugin\CheckoutDoSaveOrderInterface> $saveOrderStack
+     *
+     * @return \Generated\Shared\Transfer\SaveOrderTransfer
+     * @see \SprykerTest\Shared\Sales\Helper\SalesDataHelper::haveOrderUsingPreparedQuoteTransfer()
+     */
+    public function haveOrderUsingPreparedQuoteTransfer(\Generated\Shared\Transfer\QuoteTransfer $quoteTransfer, ?string $stateMachineProcessName = NULL, array $saveOrderStack = []): \Generated\Shared\Transfer\SaveOrderTransfer {
+        return $this->getScenario()->runStep(new \Codeception\Step\Action('haveOrderUsingPreparedQuoteTransfer', func_get_args()));
     }
 }
